refactor: add `src` folder
re #55
