+++ src/tool.js
